+++ CHANGELOG.md
@@ -8,3 +8,4 @@ All notable changes to the "fed-ponavljanje-za-ispit-deprecated" extension will 
 - Add webview and make it searchable (no content)
 - add markdown view with content
 - changed VSCode compatibility - compatible with v1.53+
+- removed test and webView (webView may come back if I find time/motivation)